requirements.txt
Flask==2.0.3
Flask-MySQLdb==1.0.1
numpy
nltk
mysql.connector
tensorflow==2.10.0
random
json
pickle
